Removed entities from team on despawn, preventing a bug regarding too many packets sent to the client, effectively banning the players; Added some mob counters on Demiurge and Scarlet Devil; Extended the time to 20min per wave; Boss spawning only shows the closest to each player + Ritual doesn't happen until every Boss is defeated

+++ src/main/java/com/github/timepsilon/comet/commands/BossList.java
@@ -32,7 +32,7 @@ public class BossList implements CommandExecutor {
             }
             for (Boss boss : bossList) {
                 Location loc = boss.getEntity().getLocation();
-                String dist = (p.getWorld().equals(loc.getWorld())) ? ""+loc.distance(p.getLocation()) : "NaN";
+                String dist = (p.getWorld().equals(loc.getWorld())) ? ""+(int)loc.distance(p.getLocation()) : "NaN";
 
                 p.sendMessage(ChatColor.YELLOW + "⚔ : " + boss.getEntity().getCustomName()
                         + ChatColor.YELLOW + ", ♥ : " + ChatColor.RED + (int)((LivingEntity)boss.getEntity()).getHealth()
